Scorm-generator-from-docs/policies-training-content-poc/requirements.txt
python-pptx
PyPDF2
gtts
pyttsx3
ffmpeg-python
Pillow
requests
google-generativeai
pydub
scormcloud
